(Somewhat) rough draft
Everything works, but it's not very pretty. I've found that I'm a lot better at the JavaScript part of this project than the CSS. I'll work on the CSS a little bit more, and I need to validate the JavaScript. I put all of the helper function declarations in the window onload main function so I don't have to use global variables, but that's probably not the best way to deal with that problem.